Introduce the SigLevel "BogusRedirect"
Pacman assumes that it encountered a dynamic mirror when it receives an
HTTP redirect. This is correct most of the time, but it is wrong e.g. when
retrieving from Bintray (which is used for the Git for Windows
repositories).

Let's support that use case by adding the `BogusRedirect` keyword to the
SigLevel specification; That way, users can specify in their
/etc/pacman.conf that certain servers redirect to completely bogus URLs
(Bintray redirects to a URL whose file name is obviously a hexadecimal
checksum, also appending HTTP parameters to the URL). In this case, Pacman
would be forced to use the correct URL for the signatures again.

Assisted-by: nalla <nalla@hamal.uberspace.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>